Separates form from list of tests
