+++ src/components/Main.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { Switch, Route } from 'react-router-dom';
 import Home from './Home';
 import Tests from './Tests';
+import RequestTests from './RequestTests';
 
 
 export default class Main extends Component {
@@ -12,6 +13,7 @@ export default class Main extends Component {
             <Switch>
               <Route exact path='/' component={Home}/>
               <Route path='/tests' component={Tests}/>
+              <Route path='/request-tests' component={RequestTests}/>
               </Switch>
           </div>
